requirement.txt
appdirs==1.4.0
configparser==3.5.0
packaging==16.8
py==1.4.32
pyparsing==2.1.10
pytest==3.0.6
PyYAML==3.12
six==1.10.0
